bsdinstall: better support custom pkgbase repo config
Currently if the user attempts to set a custom repo configuration with
$BSDINSTALL_PKG_REPOS_DIR as documented in bsdinstall(8) they will still
be prompted if they would like to perform an offline or network
installation (assuming offline packages are available).

Furthermore, if they select an offline installation, bsdinstall will
silently overwrite the BSDINSTALL_PKG_REPOS_DIR variable before invoking
the pkgbase target.

This patch fixes that issue and improves the user experience when
$BSDINSTALL_PKG_REPOS_DIR is set. There is no change in behavior if
$BSDINSTALL_PKG_REPOS_DIR is not set.

Sponsored by:	The FreeBSD Foundation